gds_design/requirements.txt
cffi==1.11.5
chardet==3.0.4
colorama==0.4.1
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.5
django-allauth==0.38.0
django-appconf==1.0.2
django-ckeditor==5.6.1
django-compressor==2.3
django-crispy-forms==1.7.2
django-extensions==2.2.3
django-js-asset==1.1.0
django-sendfile==0.3.11
et-xmlfile==1.0.1
gunicorn==19.9.0
hashids==1.2.0
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.15.1
mysqlclient==1.4.4
oauthlib==3.0.0
openpyxl==3.0.0
parso==0.5.1
pickleshare==0.7.5
Pillow==6.2.0
prompt-toolkit==2.0.10
pycparser==2.19
Pygments==2.4.2
python3-openid==3.1.0
pytz==2018.9
rcssmin==1.0.6
requests==2.21.0
requests-oauthlib==1.2.0
rjsmin==1.1.0
sentry-sdk==0.12.3
six==1.12.0
traitlets==4.3.3
urllib3==1.24.1
wcwidth==0.1.7
